+++ wsb-jpa1/src/main/java/com/capgemini/wsb/dto/PatientTO.java
@@ -17,7 +17,7 @@ public class PatientTO implements Serializable
     private Boolean isEmployed;
 
     private AddressTO address;
-    private List<Integer> visitsById;
+    private List<Long> visitsById;
 
     public Long getId() {
         return id;
@@ -91,11 +91,11 @@ public class PatientTO implements Serializable
         this.address = address;
     }
 
-    public List<Integer> getVisitsById() {
+    public List<Long> getVisitsById() {
         return visitsById;
     }
 
-    public void setVisitsById(List<Integer> visitsById) {
+    public void setVisitsById(List<Long> visitsById) {
         this.visitsById = visitsById;
     }
